finish decompiling field_control_avatar.c (#182)
* start decompiling field_control_avatar.c

* remove field_control_player_avatar.c

* finish decompiling

* cleanup

* var and flag constants

+++ src/field_poison.c
@@ -106,7 +106,7 @@ void ExecuteWhiteOut(void)
     ScriptContext1_Stop();
 }
 
-u8 overworld_poison(void)
+s32 overworld_poison(void)
 {
     struct Pokemon *pkmn = &gPlayerParty[0];
     u32 numPoisoned = 0;
